refactor(dummyData): rename folder
